Derive From impl for app state
