+++ kanin/src/tests/basic.rs
@@ -1,10 +1,8 @@
-use std::sync::Mutex;
+use std::sync::{Arc, Mutex};
 
 use lapin::{message::Delivery, Channel};
 
-use crate::{error::FromError, extract::State};
-
-use super::*;
+use crate::{error::FromError, extract::State, App, AppState, HandlerError, Respond};
 
 #[derive(Debug)]
 struct MyResponse(String);
@@ -40,7 +38,7 @@ async fn handler_with_two_extractors(_channel: Channel, _delivery: Delivery) -> 
     MyResponse("hello".into())
 }
 
-async fn handler_with_state_extractor(state: State<Mutex<u32>>) -> MyResponse {
+async fn handler_with_state_extractor(state: State<Arc<Mutex<u32>>>) -> MyResponse {
     let mut request_count = state.lock().unwrap();
     *request_count += 1;
 
@@ -48,22 +46,30 @@ async fn handler_with_state_extractor(state: State<Mutex<u32>>) -> MyResponse {
 }
 
 /// A handler that doesn't respond just doesn't return anything.
-async fn listener(state: State<Mutex<u32>>) {
+async fn listener(state: State<Arc<Mutex<u32>>>) {
     let mut request_count = state.lock().unwrap();
     // We just care about changing the state here, we don't want to reply with anything.
     *request_count += 1;
 }
 
+#[derive(AppState)]
+struct MyAppState(Arc<Mutex<u32>>);
+
+// This also works.
+#[allow(dead_code)]
+#[derive(AppState)]
+struct MyNamedAppState {
+    my_state: Arc<Mutex<u32>>,
+}
+
 /// At the moment, this just verifies that the above handlers compile and work as handlers.
 #[tokio::test]
 async fn it_compiles() {
-    let _ignore = App::new()
+    let _ignore = App::new(MyAppState(Arc::new(Mutex::new(187))))
         .handler("routing_key_0", handler)
         .handler("routing_key_1", handler_with_channel)
         .handler("routing_key_2", handler_with_delivery)
         .handler("routing_key_3", handler_with_two_extractors)
         .handler("routing_key_4", handler_with_state_extractor)
-        .handler("routing_key_5", listener)
-        .run("amqp_address")
-        .await;
+        .handler("routing_key_5", listener);
 }
